Add files via upload
This is a Python project which scrapes news. The interface is done using tkinter. It has 2 sections: Headlines and detailed new. If we choose headlines the program scrapes 3 news headlines of 5 different categories from NDTV website using selenium and BeautifulSoup. We are given option to either view the news in PDF or document format. On choosing detailed news, we can input a number and the application scrapes the inputed number of news in detail. APIs are used to get the news. Again we have choices to view it in PDF or document format.